apps/inventory/pipeline/requirements.txt
# Inventory Pipeline dependencies
# Removed: gspread, google-auth, google-auth-oauthlib, google-cloud-bigquery,
#           prophet, cmdstanpy, pystan

requests>=2.31.0
python-dotenv>=1.0.0
numpy>=1.26.0
pandas>=2.2.0
anthropic>=0.28.0
ShopifyAPI>=12.5.0
python-amazon-sp-api>=0.19.0
slack-sdk>=3.27.0
sendgrid>=6.11.0
